+++ mithril-client/src/file_downloader/mock_builder.rs
@@ -1,9 +1,14 @@
 use anyhow::anyhow;
 use mockall::predicate;
+use std::collections::BTreeMap;
+use std::fs::File;
 use std::path::{Path, PathBuf};
 
+use sha2::{Digest, Sha256};
+
 use mithril_common::{
-    entities::{CompressionAlgorithm, FileUri},
+    crypto_helper::ManifestSigner,
+    entities::{AncillaryFilesManifest, CompressionAlgorithm, FileUri},
     StdResult,
 };
 
@@ -29,6 +34,18 @@ pub struct MockFileDownloaderBuilder {
     param_target_dir: Option<PathBuf>,
     param_compression_algorithm: Option<Option<CompressionAlgorithm>>,
     returning_func: Option<MockFileDownloaderBuilderReturningFunc>,
+}
+
+/// Parameters to create fake ancillary files with a (potentially) signed manifest when calling
+/// [MockFileDownloaderBuilder::with_success_and_create_fake_ancillary_files]
+#[derive(bon::Builder)]
+pub struct FakeAncillaryFileBuilder {
+    /// List of ancillary files to create and include in the manifest (created files will be empty)
+    files_in_manifest_to_create: Vec<String>,
+    /// List of files to create and not to include in the manifest (created files will be empty)
+    files_not_in_manifest_to_create: Option<Vec<String>>,
+    /// The signer to use to sign the manifest, manifest will not be signed if None
+    sign_manifest: Option<ManifestSigner>,
 }
 
 impl Default for MockFileDownloaderBuilder {
@@ -62,6 +79,52 @@ impl MockFileDownloaderBuilder {
     pub fn with_failure(self) -> Self {
         self.with_returning(Box::new(|_, _, _, _, _| {
             Err(anyhow!("Download unpack failed"))
+        }))
+    }
+
+    /// The MockFileDownloader will succeed and will create ancillary files and a manifest
+    /// in the target directory (target directory must exist otherwise the function will panic).
+    pub fn with_success_and_create_fake_ancillary_files(
+        self,
+        fake_ancillary_file_builder: FakeAncillaryFileBuilder,
+    ) -> Self {
+        self.with_returning(Box::new(move |_, _, target_dir: &Path, _, _| {
+            for filename in [
+                fake_ancillary_file_builder
+                    .files_in_manifest_to_create
+                    .as_slice(),
+                fake_ancillary_file_builder
+                    .files_not_in_manifest_to_create
+                    .as_ref()
+                    .unwrap_or(&vec![])
+                    .as_slice(),
+            ]
+            .concat()
+            {
+                File::create(target_dir.join(filename)).unwrap();
+            }
+
+            let empty_file_sha256 = hex::encode(Sha256::digest(""));
+            let mut ancillary_manifest = AncillaryFilesManifest {
+                data: BTreeMap::from_iter(
+                    fake_ancillary_file_builder
+                        .files_in_manifest_to_create
+                        .iter()
+                        .map(|filename| (PathBuf::from(filename), empty_file_sha256.clone())),
+                ),
+                signature: None,
+            };
+
+            ancillary_manifest.signature = fake_ancillary_file_builder
+                .sign_manifest
+                .as_ref()
+                .map(|signer| signer.sign(&ancillary_manifest.compute_hash()));
+
+            let ancillary_file =
+                File::create(target_dir.join(AncillaryFilesManifest::ANCILLARY_MANIFEST_FILE_NAME))
+                    .unwrap();
+            serde_json::to_writer(ancillary_file, &ancillary_manifest).unwrap();
+            Ok(())
         }))
     }
 
@@ -152,5 +215,71 @@ impl MockFileDownloaderBuilder {
         let mock_file_downloader = self.build();
 
         Self::from_mock(mock_file_downloader)
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use mithril_common::temp_dir_create;
+
+    use crate::file_downloader::FileDownloader;
+
+    use super::*;
+
+    #[tokio::test]
+    async fn can_be_parametrized_to_create_fake_ancillary_files_with_signed_manifest() {
+        let target_dir = temp_dir_create!();
+        let ancillary_signer = ManifestSigner::create_deterministic_signer();
+        let verifier = ancillary_signer.verification_key();
+        let fake_ancillary_file_builder = FakeAncillaryFileBuilder::builder()
+            .files_in_manifest_to_create(vec!["file1.txt".to_string(), "file2.txt".to_string()])
+            .files_not_in_manifest_to_create(vec!["not_in_manifest.txt".to_string()])
+            .sign_manifest(ancillary_signer)
+            .build();
+
+        let mock_file_downloader = MockFileDownloaderBuilder::default()
+            .with_file_uri("http://ancillary.tar.gz")
+            .with_success_and_create_fake_ancillary_files(fake_ancillary_file_builder)
+            .build();
+
+        mock_file_downloader
+            .download_unpack(
+                &FileDownloaderUri::FileUri(FileUri("http://ancillary.tar.gz".to_string())),
+                0,
+                &target_dir,
+                Some(CompressionAlgorithm::default()),
+                DownloadEvent::Ancillary {
+                    download_id: "whatever".to_string(),
+                },
+            )
+            .await
+            .unwrap();
+
+        let ancillary_manifest: AncillaryFilesManifest = serde_json::from_reader(
+            File::open(target_dir.join(AncillaryFilesManifest::ANCILLARY_MANIFEST_FILE_NAME))
+                .unwrap(),
+        )
+        .unwrap();
+
+        assert!(target_dir.join("file1.txt").exists());
+        assert!(target_dir.join("file2.txt").exists());
+        assert!(target_dir.join("not_in_manifest.txt").exists());
+        const EMPTY_SHA256_HASH: &str =
+            "e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855";
+        assert_eq!(
+            ancillary_manifest.data,
+            BTreeMap::from([
+                (PathBuf::from("file1.txt"), EMPTY_SHA256_HASH.to_string()),
+                (PathBuf::from("file2.txt"), EMPTY_SHA256_HASH.to_string()),
+            ])
+        );
+        verifier
+            .verify(
+                &ancillary_manifest.compute_hash(),
+                &ancillary_manifest
+                    .signature
+                    .expect("Manifest should be signed"),
+            )
+            .expect("Signed manifest should have a valid signature");
     }
 }