+++ CHANGELOG.md
@@ -35,6 +35,7 @@ As a minor extension, we have adopted a slightly different versioning convention
 
   - Creation of two separate archives for the immutable files and for the ancillary files.
   - Added a signed manifest file to the ancillary archive (contains the list of all files in the archive and their sha256 hashes).
+  - Added client validation of the signature of the manifest file and the integrity of the files in the archive after downloading an ancillary archive.
 
 - **UNSTABLE** Implement a follower signer registration mode in the aggregator.
 
@@ -45,6 +46,7 @@ As a minor extension, we have adopted a slightly different versioning convention
   - Implement an example crate for the signed entity type `CardanoDatabase`.
   - Lighter ancillary archive by only including what's strictly necessary: the latest ledger file and the last immutable file trio.
   - Added a signed manifest file to the ancillary archive (contains the list of all files in the archive and their sha256 hashes).
+  - Added client validation of the signature of the manifest file and the integrity of the files in the archive after downloading an ancillary archive.
 
 - Crates versions: